update models & states
