+++ src/models/diary.ts
@@ -1,0 +1,10 @@
+import { breakChar } from "@/utils";
+
+export type DiaryText = string[];
+export const convertTextToDiary = (rawText: string): DiaryText => {
+    return rawText.split(breakChar);
+};
+
+export const convertPosToIndex = (pos: number, diary: DiaryText): number => {
+    return diary.findIndex((str, idx) => diary.slice(0, idx + 1).join("").length >= pos);
+};
