+++ src/models/index.ts
@@ -1,3 +1,4 @@
+export * from "./diary";
 export * from "./experience";
 export * from "./mutate";
 export * from "./receiver";
