+++ src/models/mutate.ts
@@ -1,8 +1,10 @@
+import { DiaryText } from "./diary";
+
 export const mutationStateList = ["ready", "pending", "cancel"] as const;
 
-export type DiaryText = string[];
 export type MutationStage = (typeof mutationStateList)[number];
 export type MutationState = {
+    diary: DiaryText;
     stage: MutationStage;
     mutatedLength: number;
 };
